requirements.txt
streamlit
pandas
numpy
gspread
oauth2client
matplotlib
pydeck
pytz
openpyxl
altair
streamlit-option-menu
paho-mqtt
